rename Eval to EvalScala
